server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
opencv-python==4.10.0.84
numpy==1.26.4
scikit-image==0.24.0
google-genai==0.2.2
easyocr==1.7.1

